refactor: move stamps API to debug API
