+++ pkg/debugapi/chequebook.go
@@ -20,7 +20,7 @@ import (
 	"github.com/gorilla/mux"
 )
 
-var (
+const (
 	errChequebookBalance           = "cannot get chequebook balance"
 	errChequebookNoAmount          = "did not specify amount"
 	errChequebookNoWithdraw        = "cannot withdraw"
@@ -35,8 +35,9 @@ var (
 	errBadGasPrice                 = "bad gas price"
 	errBadGasLimit                 = "bad gas limit"
 
-	gasPriceHeader = "Gas-Price"
-	gasLimitHeader = "Gas-Limit"
+	gasPriceHeader  = "Gas-Price"
+	gasLimitHeader  = "Gas-Limit"
+	immutableHeader = "Immutable"
 )
 
 type chequebookBalanceResponse struct {
